Base Summary Function

This function does the actual calculations for summarize_stats_CSV. It calculates mean and quantiles, writing output to the appropriate folder.

base_MQ(
  fList,
  oDir,
  sName,
  nodeNames,
  nNodes,
  genos,
  nGenos,
  times,
  nTimes,
  num_repss,
  mean,
  quantiles,
  oDepth
)

Arguments

fList
File list, all files for this stage, organized by repetition

oDir
Output directory

sName
Stage signifier

nodeNames
Properly formatted vector of node names for printing

nNodes
Number of nodes in the simulation

genos
Vector of genotypes for the header

nGenos
Number of genotypes

times
Vector of sampling times

nTimes
Number of sampled times

num_repss
Number of repetitions from the simulation

mean
Boolean, calculate mean or not

quantiles
Vector of quantiles to calculate, or NULL

oDepth
Max(1, number of quantiles)

Value

None