Base Aquatic Function for Erlang-Stage Summary

This function takes a given aquatic (egg, larval, pupal) stage and sums over the genotypes, returning summary trajectories by Erlang-distributed stage.

base_aquatic_stage(out, spn_P, elp)

Arguments

out
the output of sim_trajectory_R

spn_P
the places of the SPN, see details

elp
stage to summarize, one of: "egg", "larvae", "pupae"

Value

a 3 to 5 column dataframe for plotting with ggplot2

Details

This function is the base function for summarize_eggs_stage, summarize_larvae_stage, and summarize_pupae_stage.

The places (spn_P) object is generated from one of the following: spn_P_lifecycle_node, spn_P_lifecycle_network, spn_P_epiSIS_node, spn_P_epiSIS_network, spn_P_epiSEIR_node, or spn_P_epiSEIR_network.

The return object depends on the data provided. If the simulation was only 1 node, then no node designation is returned. If only one repetition was performed, no rep designation is returned. Columns always returned include: time, Erlang-stage, and value.