Calculate Equilibrium for Mosquito SEI - Human Imperial Model

In decoupled sampling, human states are handled separately from mosquito states. The function equilibrium_Imperial_decoupled_human calculates the distribution of humans at equilibrium required for the Imperial model of malaria transmission. Here we use parameters from that model to calculate the equilibrium states of Susceptible-Exposed-Infectious (SEI) female mosquitoes

Usage

equilibrium_SEI_Imperial(
  params,
  node_list = "b",
  NF = NULL,
  phi = 0.5,
  NH = NULL,
  log_dd = TRUE,
  spn_P,
  pop_ratio_Aq = NULL,
  pop_ratio_F = NULL,
  pop_ratio_M = NULL,
  pop_ratio_H = 1,
  cube
)

Arguments

params

a named list of parameters (see details)

node_list

list of geospatial nodes

NF

number of female mosquitoes

phi

sex ratio of mosquitoes at emergence

NH

vector of humans at equilibrium

log_dd

Boolean: TRUE implies logistic density dependence, FALSE implies Lotka-Volterra model

spn_P

the set of places (P) (see details)

pop_ratio_Aq

May be empty; if not, a named vector or matrix. (see details)

pop_ratio_F

May be empty; if not, a named vector or matrix. (see details)

pop_ratio_M

May be empty; if not, a named vector or matrix. (see details)

pop_ratio_H

Prevalence in human-only nodes

cube

an inheritance cube from the MGDrivE package (e.g. cubeMendelian)

Value

a vector of the equilibrium number of females in each SEI stage

Details

Imperial model sampling is currently only supported for one-node dynamics: a single node with mosquitoes parameterized by the distribution of human states. These nodes are set using the node_list parameter. Mosquito-only node equilibrium calls equilibrium_lifeycle, which follows one of two models: classic logistic dynamics or the Lotka-Volterra competition model. This is determined by the parameter log_dd, and it changes elements of the return list: K is returned for logistic dynamics, or gamma is returned for Lotka-Volterra dynamics. This is parameterized with the NF parameter to define the adult female numbers. This parameter only needs to be supplied if there are mosquito-only nodes.

For human and mosquito nodes, this function calculates the number of SEI mosquitoes in each state.

The places (spn_P) object is generated from one of the following: spn_P_lifecycle_node, spn_P_lifecycle_network, spn_P_epiSIS_node, spn_P_epiSIS_network, spn_P_epiSEIR_node, or spn_P_epiSEIR_network.

The initial population genotype ratios are set by supplying the pop_ratio_Aq, pop_ratio_F, and pop_ratio_M values. The default value is NULL, and the function will use the wild-type alleles provided in the cube object. However, one can supply several different objects to set the initial genotype ratios. All genotypes provided must exist in the cube (this is checked by the function). If a single, named vector is provided, then all patches will be initialized with the same ratios. If a matrix is provided, with the number of columns (and column names) giving the initial genotypes, and a row for each patch, each patch can be set to a different initial ratio. The three parameters do not need to match each other.

The params argument supplies all of the ecological and epidemiological parameters necessary to calculate equilibrium values. This is used to set the initial population distribution and during the simulation to maintain equilibrium. This params must include the following named parameters, noted as being the same as lifecycle parameters, or new for the epidemiological equilibrium

(Lifecycle parameters)

qE: inverse of mean duration of egg stage

nE: shape parameter of Erlang-distributed egg stage

qL: inverse of mean duration of larval stage

nL: shape parameter of Erlang-distributed larval stage

qP: inverse of mean duration of pupal stage

nP: shape parameter of Erlang-distributed pupal stage

muE: egg mortality

muL: density-independent larvae mortality

muP: pupae mortality

muF: adult female mortality, supplied from Imperial equilibrium function

muM: adult male mortality, supplied from Imperial equilibrium function

beta: egg-laying rate, daily

nu: mating rate of unmated females

(Epidemiological parameters)

NH: number of humans, can be a vector

FOIv: force of infection on mosquitoes, supplied from Imperial equilibrium function

Iv_eq: per-capita proportion of infectious mosquitoes The return list contains all of the parameters necessary later in the simulations.