Model Parameter List Creation

model_param_list_create creates list of model parameters to be used within equilibrium_init_create

Usage

imperial_model_param_list_create(
  eta = 1/(21 * 365),
  rho = 0.85,
  a0 = 2920,
  sigma2 = 1.67,
  max_age = 100 * 365,
  rA = 1/195,
  rT = 0.2,
  rD = 0.2,
  rU = 1/110.299,
  rP = 1/15,
  dE = 12,
  delayGam = 12.5,
  cD = 0.0676909,
  cT = 0.322 * cD,
  cU = 0.006203,
  gamma1 = 1.82425,
  d1 = 0.160527,
  dID = 3650,
  ID0 = 1.577533,
  kD = 0.476614,
  uD = 9.44512,
  aD = 8001.99,
  fD0 = 0.007055,
  gammaD = 4.8183,
  alphaA = 0.75735,
  alphaU = 0.185624,
  b0 = 0.590076,
  b1 = 0.5,
  dB = 3650,
  IB0 = 43.8787,
  kB = 2.15506,
  uB = 7.19919,
  theta0 = 0.0749886,
  theta1 = 0.0001191,
  iv0 = 1.09629,
  kv = 2.00048,
  av = 2493.41,
  gammaV = 2.91282,
  fvS = 0.141195,
  pctMort = 0.215,
  phi0 = 0.791666,
  phi1 = 0.000737,
  dCA = 10950,
  IC0 = 18.02366,
  kC = 2.36949,
  uCA = 6.06349,
  PM = 0.774368,
  dCM = 67.6952,
  dVM = 76.8365,
  dVA = 30 * 365,
  PVM = 0.195768,
  uVA = 11.4321,
  tau1 = 0.69,
  tau2 = 2.31,
  muF = 0.132,
  nEIP = 3,
  qEIP = 1/10,
  Q0 = 0.92,
  DY = 365,
  thetaB = 0.89,
  thetaI = 0.97,
  r_llin = 0.56,
  s_llin = 0.03,
  r_irs = 0.6,
  s_irs = 0,
  qE = 1/3,
  nE = 2,
  qL = 1/7,
  nL = 3,
  qP = 1/1,
  nP = 2,
  muE = 0.05,
  muL = 0.15,
  muP = 0.05,
  muM = 0.132,
  eps = 58.9,
  nu = 1/(4/24),
  NH = 1000,
  ...
)

Arguments

eta

Death rate for expoential population distribtuion, i.e. 1/Mean Population Age. Default = 0.0001305

rho

Age-dependent biting parameter. Default = 0.85

a0

Age-dependent biting parameter. Default = 2920

sigma2

Variance of the log heterogeneity in biting rates. Default = 1.67

max_age

Maximum age in days. Default = 100*365

rA

Rate of leaving asymptomatic infection. Default = 0.00512821

rT

Rate of leaving treatment. Default = 0.2

rD

Rate of leaving clinical disease. Default = 0.2

rU

Rate of recovering from subpatent infection. Default = 0.00906627

rP

Rate of leaving prophylaxis. Default = 0.06666667

dE

Latent period of human infection. Default = 12

delayGam

Lag from parasites to infectious gametocytes. Default = 12.5

cD

Untreated disease contribution to infectiousness. Default = 0.0676909

cT

Treated disease contribution to infectiousness. Default = 0.322 * cD

cU

Subpatent disease contribution to infectiousness. Default = 0.006203

gamma1

Parameter for infectiousness of state A. Default = 1.82425

d1

Minimum probability due to maximum immunity. Default = 0.160527

dID

Inverse of decay rate. Default = 3650

ID0

Scale parameter. Default = 1.577533

kD

Shape parameter. Default = 0.476614

uD

Duration in which immunity is not boosted. Default = 9.44512

aD

Scale parameter relating age to immunity. Default = 8001.99

fD0

Time-scale at which immunity changes with age. Default = 0.007055

gammaD

Shape parameter relating age to immunity. Default = 4.8183

alphaA

PCR detection probability parameters state A. Default = 0.757

alphaU

PCR detection probability parameters state U. Default = 0.186

b0

Maximum probability due to no immunity. Default = 0.590076

b1

Maximum relative reduction due to immunity. Default = 0.5

dB

Inverse of decay rate. Default = 3650

IB0

Scale parameter. Default = 43.8787

kB

Shape parameter. Default = 2.15506

uB

Duration in which immunity is not boosted. Default = 7.19919

theta0

Maximum probability of severe infection due to no immunity. Default = 0.0749886

theta1

Maximum reduction due to to immunity. Default = 0.0001191

iv0

Scale parameter. Default = 1.09629

kv

Shape parameter. Default = 2.00048

av

Age-dependent modifier. Default = 2493.41

gammaV

Age-dependent modifier. Default = 2.91282

fvS

Age-dependent modifier. Default = 0.141195

pctMort

Percentage of severe cases that die. Default = 0.215

phi0

Maximum probability due to no immunity. Default = 0.791666

phi1

Maximum relative reduction due to immunity. Default = 0.000737

dCA

Inverse of decay rate. Default = 10950

IC0

Scale parameter. Default = 18.02366

kC

Shape parameter. Default = 2.36949

uCA

Duration in which immunity is not boosted. Default = 6.06349

PM

New-born immunity relative to mother’s. Default = 0.774368

dCM

Inverse of decay rate of maternal immunity. Default = 67.6952

dVM

Inverse of decay rate. Default = 76.8365

dVA

Inverse of decay rate. Default = 30 * 365

PVM

New-born immunity to severe disease relative to mothers. Default = 0.195768

uVA

Duration in which immunity to severe disease is not boosted. Default = 11.4321

tau1

Duration of host seeking, assumed to be constant between species. Default = 0.69

tau2

Duration of mosquito resting after feed. Default = 2.31

muF

Daily mortality of adult mosquitos. Default = 0.132

nEIP

Number of Erlang-distributed EIP compartments. Default = 6

qEIP

Inverse of the mean duration of the EIP. Default = 1/10 (days)

Q0

Anthrophagy probability. Default = 0.92

DY

number of days in a year. Default = 365

thetaB

proportion of bites on a person in bed. Default = 0.89

thetaI

proportion of bites on a person outdoors. Default = 0.97

r_llin

probability of repeating a feeding attempt due to LLINs. Default = 0.56

s_llin

probability of feeding and surviving in presence of LLINs. Default = 0.03

r_irs

probability of repeating a feeding attempt due to IRS. Default = 0.60

s_irs

probability of feeding and surviving in presence of IRS. Default = 0

qE

mosquito egg lifecycle parameter. Default = 1/3

nE

mosquito egg lifecycle parameter. Default = 2

qL

mosquito larval lifecycle parameter. Default = 1/7

nL

mosquito larval lifecycle parameter. Default = 3

qP

mosquito pupae lifecycle parameter. Default = 1/1

nP

mosquito pupae lifecycle parameter. Default = 2

muE

death rate of egg stage. Default = 0.05

muL

death rate of larval stage. Default = 0.15

muP

death rate of pupae stage. Default = 0.05

muM

death rate of male adult stage. Default = 0.132

eps

eggs laid per day. Default = 58.9

nu

mosquito lifecycle parameter. Default = 1/(4/24

NH

number of humans. Default = 1000

...

Any other parameters needed for non-standard model. If they share the same name as any of the defined parameters model_param_list_create will stop. You can either write any extra parameters you like individually, e.g. model_param_list_create(extra1 = 1, extra2 = 2) and these parameteres will appear appended to the returned list, or you can pass explicitly the ellipsis argument as a list created before, e.g. model_param_list_create(...=list(extra1 = 1, extra2 = 2))

Value

A named vector of all baseline parameters required by the Imperial malaria model.

This function creates all of the necessary parameters for the Imperial model. Parameters furnished by MGDrivE will be removed from this function. Adapted from: https://github.com/mrc-ide/deterministic-malaria-model/blob/master/R/model_parameters.R

A newer version of the model also includes parameters for severe disease. See: https://github.com/mrc-ide/malariasimulation for details.

Examples

imperial_model_param_list_create(NH=1500)
#> $DY
#> [1] 365
#> 
#> $eta
#> [1] 0.0001304631
#> 
#> $rho
#> [1] 0.85
#> 
#> $a0
#> [1] 2920
#> 
#> $sigma2
#> [1] 1.67
#> 
#> $max_age
#> [1] 36500
#> 
#> $rA
#> [1] 0.005128205
#> 
#> $rT
#> [1] 0.2
#> 
#> $rD
#> [1] 0.2
#> 
#> $rU
#> [1] 0.009066265
#> 
#> $rP
#> [1] 0.06666667
#> 
#> $dE
#> [1] 12
#> 
#> $delayGam
#> [1] 12.5
#> 
#> $cD
#> [1] 0.0676909
#> 
#> $cT
#> [1] 0.02179647
#> 
#> $cU
#> [1] 0.006203
#> 
#> $gamma1
#> [1] 1.82425
#> 
#> $d1
#> [1] 0.160527
#> 
#> $dID
#> [1] 3650
#> 
#> $ID0
#> [1] 1.577533
#> 
#> $kD
#> [1] 0.476614
#> 
#> $uD
#> [1] 9.44512
#> 
#> $aD
#> [1] 8001.99
#> 
#> $fD0
#> [1] 0.007055
#> 
#> $gammaD
#> [1] 4.8183
#> 
#> $alphaA
#> [1] 0.75735
#> 
#> $alphaU
#> [1] 0.185624
#> 
#> $b0
#> [1] 0.590076
#> 
#> $b1
#> [1] 0.5
#> 
#> $dB
#> [1] 3650
#> 
#> $IB0
#> [1] 43.8787
#> 
#> $kB
#> [1] 2.15506
#> 
#> $uB
#> [1] 7.19919
#> 
#> $phi0
#> [1] 0.791666
#> 
#> $phi1
#> [1] 0.000737
#> 
#> $dCA
#> [1] 10950
#> 
#> $IC0
#> [1] 18.02366
#> 
#> $kC
#> [1] 2.36949
#> 
#> $uCA
#> [1] 6.06349
#> 
#> $PM
#> [1] 0.774368
#> 
#> $dCM
#> [1] 67.6952
#> 
#> $theta0
#> [1] 0.0749886
#> 
#> $theta1
#> [1] 0.0001191
#> 
#> $iv0
#> [1] 1.09629
#> 
#> $kv
#> [1] 2.00048
#> 
#> $av
#> [1] 2493.41
#> 
#> $gammaV
#> [1] 2.91282
#> 
#> $fvS
#> [1] 0.141195
#> 
#> $pctMort
#> [1] 0.215
#> 
#> $dVM
#> [1] 76.8365
#> 
#> $dVA
#> [1] 10950
#> 
#> $PVM
#> [1] 0.195768
#> 
#> $uVA
#> [1] 11.4321
#> 
#> $qE
#> [1] 0.3333333
#> 
#> $nE
#> [1] 2
#> 
#> $qL
#> [1] 0.1428571
#> 
#> $nL
#> [1] 3
#> 
#> $qP
#> [1] 1
#> 
#> $nP
#> [1] 2
#> 
#> $muE
#> [1] 0.05
#> 
#> $muL
#> [1] 0.15
#> 
#> $muP
#> [1] 0.05
#> 
#> $muF
#> [1] 0.132
#> 
#> $muM
#> [1] 0.132
#> 
#> $eps
#> [1] 58.9
#> 
#> $nu
#> [1] 6
#> 
#> $NH
#> [1] 1500
#> 
#> $nEIP
#> [1] 3
#> 
#> $qEIP
#> [1] 0.1
#> 
#> $tau1
#> [1] 0.69
#> 
#> $tau2
#> [1] 2.31
#> 
#> $Q0
#> [1] 0.92
#> 
#> $fv0
#> [1] 0.3333333
#> 
#> $av0
#> [1] 0.3066667
#> 
#> $Surv0
#> [1] 0.334898
#> 
#> $beta
#> [1] 16.00183
#> 
#> $thetaB
#> [1] 0.89
#> 
#> $thetaI
#> [1] 0.97
#> 
#> $r_llin
#> [1] 0.56
#> 
#> $s_llin
#> [1] 0.03
#> 
#> $r_irs
#> [1] 0.6
#> 
#> $s_irs
#> [1] 0
#> 
imperial_model_param_list_create(qE=1/4)
#> $DY
#> [1] 365
#> 
#> $eta
#> [1] 0.0001304631
#> 
#> $rho
#> [1] 0.85
#> 
#> $a0
#> [1] 2920
#> 
#> $sigma2
#> [1] 1.67
#> 
#> $max_age
#> [1] 36500
#> 
#> $rA
#> [1] 0.005128205
#> 
#> $rT
#> [1] 0.2
#> 
#> $rD
#> [1] 0.2
#> 
#> $rU
#> [1] 0.009066265
#> 
#> $rP
#> [1] 0.06666667
#> 
#> $dE
#> [1] 12
#> 
#> $delayGam
#> [1] 12.5
#> 
#> $cD
#> [1] 0.0676909
#> 
#> $cT
#> [1] 0.02179647
#> 
#> $cU
#> [1] 0.006203
#> 
#> $gamma1
#> [1] 1.82425
#> 
#> $d1
#> [1] 0.160527
#> 
#> $dID
#> [1] 3650
#> 
#> $ID0
#> [1] 1.577533
#> 
#> $kD
#> [1] 0.476614
#> 
#> $uD
#> [1] 9.44512
#> 
#> $aD
#> [1] 8001.99
#> 
#> $fD0
#> [1] 0.007055
#> 
#> $gammaD
#> [1] 4.8183
#> 
#> $alphaA
#> [1] 0.75735
#> 
#> $alphaU
#> [1] 0.185624
#> 
#> $b0
#> [1] 0.590076
#> 
#> $b1
#> [1] 0.5
#> 
#> $dB
#> [1] 3650
#> 
#> $IB0
#> [1] 43.8787
#> 
#> $kB
#> [1] 2.15506
#> 
#> $uB
#> [1] 7.19919
#> 
#> $phi0
#> [1] 0.791666
#> 
#> $phi1
#> [1] 0.000737
#> 
#> $dCA
#> [1] 10950
#> 
#> $IC0
#> [1] 18.02366
#> 
#> $kC
#> [1] 2.36949
#> 
#> $uCA
#> [1] 6.06349
#> 
#> $PM
#> [1] 0.774368
#> 
#> $dCM
#> [1] 67.6952
#> 
#> $theta0
#> [1] 0.0749886
#> 
#> $theta1
#> [1] 0.0001191
#> 
#> $iv0
#> [1] 1.09629
#> 
#> $kv
#> [1] 2.00048
#> 
#> $av
#> [1] 2493.41
#> 
#> $gammaV
#> [1] 2.91282
#> 
#> $fvS
#> [1] 0.141195
#> 
#> $pctMort
#> [1] 0.215
#> 
#> $dVM
#> [1] 76.8365
#> 
#> $dVA
#> [1] 10950
#> 
#> $PVM
#> [1] 0.195768
#> 
#> $uVA
#> [1] 11.4321
#> 
#> $qE
#> [1] 0.25
#> 
#> $nE
#> [1] 2
#> 
#> $qL
#> [1] 0.1428571
#> 
#> $nL
#> [1] 3
#> 
#> $qP
#> [1] 1
#> 
#> $nP
#> [1] 2
#> 
#> $muE
#> [1] 0.05
#> 
#> $muL
#> [1] 0.15
#> 
#> $muP
#> [1] 0.05
#> 
#> $muF
#> [1] 0.132
#> 
#> $muM
#> [1] 0.132
#> 
#> $eps
#> [1] 58.9
#> 
#> $nu
#> [1] 6
#> 
#> $NH
#> [1] 1000
#> 
#> $nEIP
#> [1] 3
#> 
#> $qEIP
#> [1] 0.1
#> 
#> $tau1
#> [1] 0.69
#> 
#> $tau2
#> [1] 2.31
#> 
#> $Q0
#> [1] 0.92
#> 
#> $fv0
#> [1] 0.3333333
#> 
#> $av0
#> [1] 0.3066667
#> 
#> $Surv0
#> [1] 0.334898
#> 
#> $beta
#> [1] 16.00183
#> 
#> $thetaB
#> [1] 0.89
#> 
#> $thetaI
#> [1] 0.97
#> 
#> $r_llin
#> [1] 0.56
#> 
#> $s_llin
#> [1] 0.03
#> 
#> $r_irs
#> [1] 0.6
#> 
#> $s_irs
#> [1] 0
#>