Simulate Trajectory From one SPN Model

This is an internal function to sim_trajectory_CSV. It does the actual sampling once all of the functions have been checked and setup.

sim_trajectory_base_CSV(
  x0,
  times,
  stepFun,
  folders,
  stage,
  events0 = NULL,
  batch = NULL,
  Sout = NULL,
  verbose = TRUE
)

Arguments

x0
the initial marking of the SPN (initial state)

times
sequence of sampling times

stepFun
a sampling function

folders
vector of folders to write output

stage
vector of life-stages to print

events0
a data.frame of events (uses the same format as required in package deSolve for consistency, see events for more information)

batch
a list of batch migration events, created from batch_migration, may be set to NULL if not used

Sout
an optional matrix to track event firings

verbose
print a progress bar?

Value

no return, prints .csv files into provided folders