Make Places (P) For a Node (SEI Mosquitoes - SIS Humans)

This function makes the set of places (P) for a SPN. It is used alone if our model is a single-node metapopulation for mosquito SEI and human SIS dynamics; otherwise it is used as part of other functions to make SPN models with larger state spaces (metapopulation models, see spn_P_epiSIS_network).

spn_P_epiSIS_node(params, cube)

Arguments

params
a named list of parameters (see details)

cube
an inheritance cube from the MGDrivE package (e.g. cubeMendelian)

Value

a list with two elements: ix contains labeled indices of the places by life stage, u is the character vector of places (P)

Details

The params argument supplies all of the ecological parameters necessary to calculate equilibrium values. This function requires the nE, nL, nP, and nEIP parameters to be specified. For more details, see equilibrium_SEI_SIS

For examples of using this function, see: vignette("epi-node", package = "MGDrivE2")