Summarize Adult Females (One Node or Metapopulation Network, SEI Mosquitoes)

For MGDrivE2 simulations of mosquito epidemiological dynamics in a single node or metapopulation network, this function sums over the male mate genotype as well as EIP bins to get population trajectories of adult female mosquitoes by their genotype and (S,E,I) status.

Usage

summarize_females_epi(out, spn_P)

Arguments

out

the output of sim_trajectory_R

spn_P

the places of the SPN, see details

Value

a 4 to 6 column dataframe for plotting with ggplot2

Details

The places (spn_P) object is generated from one of the following: spn_P_epiSIS_node, spn_P_epiSIS_network, spn_P_epiSEIR_node, or spn_P_epiSEIR_network.

The return object depends on the data provided. If the simulation was only 1 node, then no node designation is returned. If only one repetition was performed, no rep designation is returned. Columns always returned include: time, inf, genotype, and value.

For examples of using this function, this or any vignette which simulates epi dynamics: vignette("epi-node", package = "MGDrivE2")