Sample Batch Migration Events

Sample batch migration events for simulation given rates of occurance and probability of destination for each patch. Batch migration can be simulated for the aquatic life stages (eggs, larvae, pupae), adult females, and/or adult males. To simulate batch migration, each life stage needs all 3 of its arguments specified. If any arguments are left unspecified (NULL), batch migration for that life stage will not be sampled. The output of this function should be passed to sim_trajectory_R or sim_trajectory_CSV as the argument batch. Calls the internal function batch_migration_stage.

batch_migration(
  SPN_P,
  tmax,
  ELPrates = NULL,
  ELPmove = NULL,
  ELPprob = NULL,
  Frates = NULL,
  Fmove = NULL,
  Fprob = NULL,
  Mrates = NULL,
  Mmove = NULL,
  Mprob = NULL
)

Arguments

SPN_P
places of the SPN

tmax
maximum time of the simulation

ELPrates
rate at which aquatic stage batch migration occurs for each node (nodes without mosquitoes should be set to NaN or NA)

ELPmove
movement matrix for destinations of aquatic stage batch migration events (diagonal will be set to zero and off-diagonal elements normalized)

ELPprob
probability for each individual to be chosen for aquatic stage batch migration events (must be same length as ELPrates)

Frates
rate at which adult female batch migration occurs for each node (nodes without mosquitoes should be set to NaN or NA)

Fmove
movement matrix for destinations of adult female batch migration events (diagonal will be set to zero and off-diagonal elements normalized)

Fprob
probability for each individual to be chosen for adult female batch migration events (must be same length as Frates)

Mrates
rate at which adult male batch migration occurs for each node (nodes without mosquitoes should be set to NaN or NA)

Mmove
movement matrix for destinations of adult male batch migration events (diagonal will be set to zero and off-diagonal elements normalized)

Mprob
probability for each individual to be chosen for adult male batch migration events (must be same length as Mrates)