Calculate Equilibrium for Lifecycle Model (Logistic or Lotka-Volterra)

This function calculates deterministic equilibria for the mosquito lifecycle model.

equilibrium_lifeycle(
  params,
  NF,
  phi = 0.5,
  log_dd = TRUE,
  spn_P,
  pop_ratio_Aq = NULL,
  pop_ratio_F = NULL,
  pop_ratio_M = NULL,
  cube
)

Arguments

params
a named list of parameters (see details)

NF
vector of female mosquitoes at equilibrium, for every population in the environment

phi
sex ratio of mosquitoes at emergence

log_dd
Boolean: TRUE implies logistic density dependence, FALSE implies Lotka-Volterra model

spn_P
the set of places (P) (see details)

pop_ratio_Aq
May be empty; if not, a named vector or matrix. (see details)

pop_ratio_F
May be empty; if not, a named vector or matrix. (see details)

pop_ratio_M
May be empty; if not, a named vector or matrix. (see details)

cube
an inheritance cube from the MGDrivE package (e.g. cubeMendelian)

Value

a list with 3 elements: init a matrix of equilibrium values for every life-cycle stage, params a list of parameters for the simulation, M0 a vector of initial conditions

Details

Equilibrium can be calculated using one of two models: classic logistic dynamics or following the Lotka-Volterra competition model. This is determined by the parameter log_dd, and it changes elements of the return list: K is returned for logistic dynamics, or gamma is returned for Lotka-Volterra dynamics.

The places (spn_P) object is generated from one of the following: spn_P_lifecycle_node, spn_P_lifecycle_network, spn_P_epiSIS_node, spn_P_epiSIS_network, spn_P_epiSEIR_node, or spn_P_epiSEIR_network.

The initial population genotype ratios are set by supplying the pop_ratio_Aq, pop_ratio_F, and pop_ratio_M values. The default value is NULL, and the function will use the wild-type alleles provided in the cube object. However, one can supply several different objects to set the initial genotype ratios. All genotypes provided must exist in the cube (this is checked by the function). If a single, named vector is provided, then all patches will be initialized with the same ratios. If a matrix is provided, with the number of columns (and column names) giving the initial genotypes, and a row for each patch, each patch can be set to a different initial ratio. The three parameters do not need to match each other.

The params argument supplies all of the ecological parameters necessary to calculate equilibrium values. This is used to set the initial population distribution and during the simulation to maintain equilibrium. params must include the following named parameters:

qE: inverse of mean duration of egg stage

nE: shape parameter of Erlang-distributed egg stage

qL: inverse of mean duration of larval stage

nL: shape parameter of Erlang-distributed larval stage

qP: inverse of mean duration of pupal stage

nP: shape parameter of Erlang-distributed pupal stage

muE: egg mortality

muL: density-independent larvae mortality

muP: pupae mortality

muF: adult female mortality

muM: adult male mortality

beta: egg-laying rate, daily

nu: mating rate of unmated females

The return list contains all of the params parameters, along with the density-dependent parameter, either K or gamma. These are the parameters necessary later in the simulations. This was done for compatibility with equilibrium_SEI_SIS, which requires several extra parameters not required further in the simulations.

For equilibrium with epidemiological parameters, see equilibrium_SEI_SIS. For equilibrium with latent humans (SEIR dynamics), see equilibrium_SEI_SEIR.