Solve for Constant Aquatic Mortality

In MGDrivE, the model was typically solved at equilibrium assuming the density-independent mortality was constant over aquatic stages (eggs, larvae, pupae), given a daily growth rate, \(r_{M}\). Given that growth rate, it solved for that mortality \(\mu_{Aqua}\) by relating it with \(R_{M}\), the per-generation growth rate of the population, calculable from \(r_{M}\) and the mean duration of life stages. This function uses uniroot to solve for \(mu_{Aqua}\).

solve_muAqua(params, rm)

Arguments

params
a named list of parameters

rm
the daily growth rate

Value

location of the root, as provided from uniroot

Details

This function needs the following parameters in params:

muF: adult female mortality

beta: rate of egg laying

phi: sex ratio at emergence

qE: inverse of mean duration of egg stage

nE: shape parameter of Erlang-distributed egg stage

qL: inverse of mean duration of larval stage

nL: shape parameter of Erlang-distributed larval stage

qP: inverse of mean duration of pupal stage

nP: shape parameter of Erlang-distributed pupal stage

Examples

theta <- list(qE = 1/4, nE = 2, qL = 1/5, nL = 3, qP = 1/6, nP = 2, muF = 1/12,
             beta = 32, phi = 0.5);
muAqatic <- solve_muAqua(params = theta, rm = 1.096)